tests/requirements.txt
mysql-connector-python
pytest
